backend/requirements.txt
numpy==1.24.3
pandas
tensorflow==2.15.0
keras>=2.2.0
keras_applications>=1.0.7
opencv-python==4.8.1.78
mtcnn>=0.1.0
h5py
efficientnet
split_folders
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
